Split v2 eval cast helpers
